Major rewrite of conditional histogram code.
Bob had the idea that it would be good to look at global vs regional
hists. V nice idea, but was tricky due to the way the data were being
loaded. I've rewritten so that the data is loaded month at a time, which
makes it relatively easy to do the gridpoint analysis (as well as doing
the old hourly analysis). This was quite a major change. Need to check
carefully that this does not alter results and check that the two give
the same answer when suitably summed.